configure test setup
