+++ src/paint.spec.js
@@ -1,0 +1,7 @@
+import paint from './paint';
+
+describe('paint', () => {
+  it('should be defined', () => {
+    expect(paint).toBeDefined();
+  });
+});
